dbのリレーションを張った

+++ db/migrate/20180429094648_create_chara_mappings.rb
@@ -1,8 +1,8 @@
 class CreateCharaMappings < ActiveRecord::Migration[5.1]
   def change
     create_table :chara_mappings do |t|
-      t.integer :image_id
-      t.integer :chara_id
+      t.bigint :image_id
+      t.bigint :chara_id
 
       t.timestamps
     end
